chore: address IntelliJ inspection warnings (safe sweep)
Mechanical clean-up driven by a multi-agent IDE-inspection survey
(workflow over 12 lib modules, 375 findings total). Restricts itself
to the categories that are safe to fix without judgment.

Three categories addressed:

1. unused-pattern-variable (15 sites): rename `__` to `_` in
   sealed-Result pattern matches. Java 22+ unnamed pattern;
   mechanical with no behavior change.
2. redundant-cast (13 → 12 sites): remove casts the compiler can
   infer from context. The (long) cast on
   DispatcherIntegrationTest.java:105 stays — it's the boxing path
   from int to Map<TopicPartition, Long>, where IntelliJ's
   "redundant" verdict was actually wrong (compile-fail without it).
3. redundant-suppression (14 sites): delete stale @SuppressWarnings
   annotations that no longer match any real warning.

Plus two unresolved-javadoc-symbol fixes (ERROR severity):
- KPipeDlqTest: qualify [Builder#withDeadLetterQueue(...)] →
  [KPipeConsumer.Builder#withDeadLetterQueue(...)]
- ParallelDispatcherTest: qualify [#close()] →
  [ParallelDispatcher#close()]

What's NOT in this PR (deferred to per-site review):

- unused-parameter (51): almost all are deliberate-ignore in test
  lambdas; needs `_` rename per-site, not a blanket fix.
- try-with-resources (39): rewriting integration tests would change
  Handle / KPipeConsumer shutdown semantics.
- busy-wait (17): intentional poll-until-condition loops in shutdown
  drain and KEY_ORDERED retry paths.
- nullable-handling (16): per-site judgment, especially AvroFormat
  registry-mode hot path.
- unused-method (11): public fluent-facade methods kept as escape
  hatches; documented in docs/escape-hatches.md.
- unchecked-assignment (9): Mockito raw-type captures and Kafka
  ConsumerRecord/ProducerRecord generics need per-site decisions.

Test plan:
- ./gradlew test  (green locally)
- ./gradlew spotlessCheck  (green)
- ./gradlew compileJava compileTestJava  (green)

+++ lib/kpipe-consumer/src/test/java/io/github/eschizoid/kpipe/consumer/KPipeBackpressureIntegrationTest.java
@@ -124,7 +124,7 @@ class KPipeBackpressureIntegrationTest {
 
       // backpressureTimeMs must be > 0
       assertTrue(
-        (long) consumer.getMetrics().get(KPipeConsumer.METRIC_BACKPRESSURE_TIME_MS) > 0,
+        consumer.getMetrics().get(KPipeConsumer.METRIC_BACKPRESSURE_TIME_MS) > 0,
         "backpressureTimeMs should be positive after a backpressure pause"
       );
 
@@ -228,7 +228,7 @@ class KPipeBackpressureIntegrationTest {
       awaitCondition(() -> mc.paused().isEmpty(), 3000);
 
       // Now it should poll the remaining records
-      awaitCondition(() -> (long) consumer.getMetrics().get("messagesReceived") == 10, 3000);
+      awaitCondition(() -> consumer.getMetrics().get("messagesReceived") == 10, 3000);
       assertEquals(10L, consumer.getMetrics().get("messagesReceived"));
 
       consumer.close();
@@ -263,7 +263,7 @@ class KPipeBackpressureIntegrationTest {
       consumer.start();
 
       // Wait for first 5 to be polled
-      awaitCondition(() -> (long) consumer.getMetrics().get("messagesReceived") == 5, 2000);
+      awaitCondition(() -> consumer.getMetrics().get("messagesReceived") == 5, 2000);
 
       // Pause manually
       consumer.pause();
@@ -284,7 +284,7 @@ class KPipeBackpressureIntegrationTest {
       consumer.resume();
 
       // Wait for the rest to be polled
-      awaitCondition(() -> (long) consumer.getMetrics().get("messagesReceived") == 10, 3000);
+      awaitCondition(() -> consumer.getMetrics().get("messagesReceived") == 10, 3000);
       assertEquals(10L, consumer.getMetrics().get("messagesReceived"));
 
       consumer.close();
@@ -341,7 +341,7 @@ class KPipeBackpressureIntegrationTest {
 
       // Assert: consumer is paused
       assertTrue(mockConsumer.paused().contains(PARTITION), "Consumer should be paused due to high lag");
-      assertTrue((long) consumer.getMetrics().get(KPipeConsumer.METRIC_BACKPRESSURE_PAUSE_COUNT) >= 1);
+      assertTrue(consumer.getMetrics().get(KPipeConsumer.METRIC_BACKPRESSURE_PAUSE_COUNT) >= 1);
 
       consumer.close();
     }
